requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib
google-auth
google-auth-oauthlib
requests
python-dotenv
google-generativeai
